refactor(ranger): Move ranger out of runtime
