+++ src/CMakeLists.txt
@@ -45,6 +45,7 @@ add_subdirectory(http)
 add_subdirectory(meta)
 add_subdirectory(nfs)
 add_subdirectory(perf_counter)
+add_subdirectory(ranger)
 add_subdirectory(redis_protocol)
 add_subdirectory(remote_cmd)
 add_subdirectory(replica)
