+++ src/aio/test/CMakeLists.txt
@@ -23,36 +23,24 @@
 # THE SOFTWARE.
 
 set(MY_PROJ_NAME dsn_aio_test)
-
-# Source files under CURRENT project directory will be automatically included.
-# You can manually set MY_PROJ_SRC to include source files under other directories.
 set(MY_PROJ_SRC "")
-
-# Search mode for source files under CURRENT project directory?
-# "GLOB_RECURSE" for recursive search
-# "GLOB" for non-recursive search
 set(MY_SRC_SEARCH_MODE "GLOB")
-
 set(MY_PROJ_LIBS
-        dsn_meta_server
-        dsn_replication_common
-        gtest
         dsn_runtime
+        dsn_meta_server
+        dsn_ranger
+        dsn_replication_common
         dsn_aio
         test_utils
+        gtest
         rocksdb
         lz4
         zstd
         snappy)
-
 set(MY_BOOST_LIBS Boost::system Boost::filesystem)
-
-# Extra files that will be installed
 set(MY_BINPLACES
-    "${CMAKE_CURRENT_SOURCE_DIR}/config.ini"
-    "${CMAKE_CURRENT_SOURCE_DIR}/clear.sh"
-    "${CMAKE_CURRENT_SOURCE_DIR}/run.sh"
-    "${CMAKE_CURRENT_SOURCE_DIR}/copy_source.txt"
-)
-
+        config.ini
+        clear.sh
+        run.sh
+        copy_source.txt)
 dsn_add_test()
