+++ src/meta/test/balancer_simulator/CMakeLists.txt
@@ -23,26 +23,14 @@
 # THE SOFTWARE.
 
 set(MY_PROJ_NAME sim_lb)
-
-# Source files under CURRENT project directory will be automatically included.
-# You can manually set MY_PROJ_SRC to include source files under other directories.
 set(MY_PROJ_SRC ../misc/misc.cpp)
-
-# Search mode for source files under CURRENT project directory?
-# "GLOB_RECURSE" for recursive search
-# "GLOB" for non-recursive search
 set(MY_SRC_SEARCH_MODE "GLOB")
-
 set(MY_PROJ_LIBS
-    dsn_meta_server
-    dsn_replication_common
-    dsn_runtime
-    hashtable
-    gtest)
-
+        dsn_runtime
+        dsn_meta_server
+        dsn_ranger
+        dsn_replication_common
+        gtest)
 set(MY_BOOST_LIBS Boost::system Boost::filesystem)
-
-# Extra files that will be installed
 set(MY_BINPLACES "")
-
 dsn_add_test()
