+++ src/ranger/CMakeLists.txt
@@ -15,24 +15,12 @@
 # specific language governing permissions and limitations
 # under the License.
 
-set(MY_PROJ_NAME dsn_replica_bulk_load_test)
+set(MY_PROJ_NAME dsn_ranger)
 set(MY_PROJ_SRC "")
 set(MY_SRC_SEARCH_MODE "GLOB")
 set(MY_PROJ_LIBS
         dsn_meta_server
-        dsn_ranger
-        dsn_replica_server
         dsn_replication_common
-        dsn_runtime
-        hashtable
-        gtest
-        test_utils
-        rocksdb
-        lz4
-        zstd
-        snappy)
-set(MY_BOOST_LIBS Boost::system Boost::filesystem)
-set(MY_BINPLACES
-        config-test.ini
-        run.sh)
-dsn_add_test()
+        dsn_utils)
+dsn_add_static_library()
+add_subdirectory(test)
