+++ src/ranger/ranger_resource_policy.cpp
@@ -17,7 +17,7 @@
 
 #include "ranger_resource_policy.h"
 
-#include "runtime/ranger/access_type.h"
+#include "ranger/access_type.h"
 #include "utils/fmt_logging.h"
 
 namespace dsn {
