+++ src/replica/storage/simple_kv/test/CMakeLists.txt
@@ -23,15 +23,11 @@
 # THE SOFTWARE.
 
 set(MY_PROJ_NAME dsn.rep_tests.simple_kv)
-
-# Search mode for source files under CURRENT project directory?
-# "GLOB_RECURSE" for recursive search
-# "GLOB" for non-recursive search
 set(MY_SRC_SEARCH_MODE "GLOB")
-
 set(MY_PROJ_LIBS
-        dsn_replica_server
         dsn_meta_server
+        dsn_ranger
+        dsn_replica_server
         dsn_replication_common
         dsn_client
         dsn.failure_detector
@@ -45,16 +41,11 @@ set(MY_PROJ_LIBS
         lz4
         zstd
         snappy)
-
 set(MY_BOOST_LIBS Boost::system Boost::filesystem)
-
-# Extra files that will be installed
 FILE(GLOB CASE_FILES "case-*")
 set(MY_BINPLACES
-    "run.sh"
-    "clear.sh"
-    "addcase.sh"
-    "${CASE_FILES}"
-)
-
+        run.sh
+        clear.sh
+        addcase.sh
+        ${CASE_FILES})
 dsn_add_test()
