+++ src/runtime/CMakeLists.txt
@@ -26,11 +26,9 @@ add_subdirectory(test)
 add_subdirectory(rpc)
 add_subdirectory(task)
 add_subdirectory(security)
-add_subdirectory(ranger)
 
 # TODO(zlw) remove perf_counter from dsn_runtime after the refactor by WuTao
 add_library(dsn_runtime STATIC
-        $<TARGET_OBJECTS:dsn_ranger>
         $<TARGET_OBJECTS:dsn.security>
         $<TARGET_OBJECTS:dsn.rpc>
         $<TARGET_OBJECTS:dsn.task>
@@ -51,8 +49,7 @@ add_library(dsn_runtime STATIC
         threadpool_code.cpp
         tool_api.cpp
         tracer.cpp
-        zlocks.cpp
-        )
+        zlocks.cpp)
 target_link_libraries(dsn_runtime PRIVATE dsn_utils sasl2 gssapi_krb5 krb5)
 define_file_basename_for_sources(dsn_runtime)
 install(TARGETS dsn_runtime DESTINATION "lib")
