+++ src/server/test/CMakeLists.txt
@@ -16,28 +16,26 @@
 # under the License.
 
 set(MY_PROJ_NAME pegasus_unit_test)
-
-set(MY_PROJ_SRC "../pegasus_server_impl.cpp"
-                "../pegasus_server_impl_init.cpp"
-                "../pegasus_manual_compact_service.cpp"
-                "../pegasus_event_listener.cpp"
-                "../pegasus_write_service.cpp"
-                "../pegasus_server_write.cpp"
-                "../capacity_unit_calculator.cpp"
-                "../pegasus_mutation_duplicator.cpp"
-                "../hotspot_partition_calculator.cpp"
-                "../meta_store.cpp"
-                "../hotkey_collector.cpp"
-                "../rocksdb_wrapper.cpp"
-                "../compaction_filter_rule.cpp"
-                "../compaction_operation.cpp"
-        )
-
+set(MY_PROJ_SRC
+        "../pegasus_server_impl.cpp"
+        "../pegasus_server_impl_init.cpp"
+        "../pegasus_manual_compact_service.cpp"
+        "../pegasus_event_listener.cpp"
+        "../pegasus_write_service.cpp"
+        "../pegasus_server_write.cpp"
+        "../capacity_unit_calculator.cpp"
+        "../pegasus_mutation_duplicator.cpp"
+        "../hotspot_partition_calculator.cpp"
+        "../meta_store.cpp"
+        "../hotkey_collector.cpp"
+        "../rocksdb_wrapper.cpp"
+        "../compaction_filter_rule.cpp"
+        "../compaction_operation.cpp")
 set(MY_SRC_SEARCH_MODE "GLOB")
-
 set(MY_PROJ_LIBS
         dsn_replica_server
         dsn_meta_server
+        dsn_ranger
         dsn_replication_common
         dsn_client
         dsn.block_service.local
@@ -54,13 +52,11 @@ set(MY_PROJ_LIBS
         pegasus_base
         gtest
         gmock
-        hashtable
-)
+        hashtable)
 add_definitions(-DPEGASUS_UNIT_TEST)
 add_definitions(-DENABLE_FAIL)
-
 set(MY_BOOST_LIBS Boost::system Boost::filesystem)
-
-set(MY_BINPLACES config.ini run.sh)
-
+set(MY_BINPLACES
+        config.ini
+        run.sh)
 dsn_add_test()
