+++ src/utils/filesystem.h
@@ -75,7 +75,9 @@ std::string remove_file_name(const std::string &path);
 
 std::string get_file_name(const std::string &path);
 
+// TODO(yingchun): The 2 functions are similar, remove concat_path_unix_style.
 std::string path_combine(const std::string &path1, const std::string &path2);
+std::string concat_path_unix_style(const std::string &prefix, const std::string &postfix);
 
 typedef std::function<int(const char *, int, struct FTW *)> ftw_handler;
 bool file_tree_walk(const std::string &dirpath, ftw_handler handler, bool recursive = true);
